restruct kube pkg
