+++ pkg/kubernetes/kube_test.go
@@ -16,33 +16,27 @@ package kube
 
 import (
 	"context"
-	"io/ioutil"
 	"path/filepath"
 	"testing"
 
 	util "github.com/keikoproj/kubedog/internal/utilities"
 	"github.com/onsi/gomega"
-	log "github.com/sirupsen/logrus"
 	appsv1 "k8s.io/api/apps/v1"
 	hpa "k8s.io/api/autoscaling/v2beta2"
 	v1 "k8s.io/api/core/v1"
 	policy "k8s.io/api/policy/v1beta1"
 	rbacv1 "k8s.io/api/rbac/v1"
-	kerrors "k8s.io/apimachinery/pkg/api/errors"
-	"k8s.io/apimachinery/pkg/api/meta"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"k8s.io/apimachinery/pkg/runtime"
-	"k8s.io/apimachinery/pkg/runtime/schema"
-	serializer "k8s.io/apimachinery/pkg/runtime/serializer/yaml"
 	fakeDiscovery "k8s.io/client-go/discovery/fake"
 	"k8s.io/client-go/dynamic"
 	fakeDynamic "k8s.io/client-go/dynamic/fake"
 	"k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/kubernetes/fake"
-	kTesting "k8s.io/client-go/testing"
 )
 
+// TODO: redistribute this test functions
 func TestPositiveNodesWithSelectorShouldBe(t *testing.T) {
 
 	var (
@@ -77,52 +71,20 @@ func TestPositiveNodesWithSelectorShouldBe(t *testing.T) {
 		},
 	})
 
-	kc := Client{
+	kc := ClientSet{
 		KubeInterface:      fakeClient,
 		DiscoveryInterface: fakeDiscovery,
 		DynamicInterface:   fakeDynamicClient,
 		FilesPath:          "../../test/templates",
 	}
 
-	err := kc.NodesWithSelectorShouldBe(1, testReadySelector, NodeStateReady)
+	err := kc.NodesWithSelectorShouldBe(1, testReadySelector, stateReady)
 	g.Expect(err).ShouldNot(gomega.HaveOccurred())
-	err = kc.NodesWithSelectorShouldBe(1, testFoundSelector, NodeStateFound)
-	g.Expect(err).ShouldNot(gomega.HaveOccurred())
-}
-
-func TestPositiveResourceOperation(t *testing.T) {
-	var (
-		err               error
-		g                 = gomega.NewWithT(t)
-		dynScheme         = runtime.NewScheme()
-		fakeDynamicClient = fakeDynamic.NewSimpleDynamicClient(dynScheme)
-		testResource      *unstructured.Unstructured
-		fakeDiscovery     = fakeDiscovery.FakeDiscovery{}
-		fakeClient        *fake.Clientset
-	)
-
-	const fileName = "test-resourcefile.yaml"
-
-	testResource, err = resourceFromYaml(fileName)
-	if err != nil {
-		t.Errorf("Failed getting the test resource from the file %v: %v", fileName, err)
-	}
-
-	fakeDiscovery.Fake = &fakeDynamicClient.Fake
-	fakeDiscovery.Resources = append(fakeDiscovery.Resources, newTestAPIResourceList(testResource.GetAPIVersion(), testResource.GetName(), testResource.GetKind()))
-
-	kc := Client{
-		KubeInterface:      fakeClient,
-		DynamicInterface:   fakeDynamicClient,
-		DiscoveryInterface: &fakeDiscovery,
-		FilesPath:          "../../test/templates",
-	}
-
-	err = kc.ResourceOperation(OperationCreate, fileName)
-	g.Expect(err).ShouldNot(gomega.HaveOccurred())
-	err = kc.ResourceOperation(OperationDelete, fileName)
+	err = kc.NodesWithSelectorShouldBe(1, testFoundSelector, stateFound)
 	g.Expect(err).ShouldNot(gomega.HaveOccurred())
 }
+
+// TODO: this doesnt seem to be testing MultipleResourcesOperation but GetMultipleResourcesFromYaml
 func TestMultipleResourcesOperation(t *testing.T) {
 
 	var (
@@ -195,187 +157,6 @@ func TestMultipleResourcesOperation(t *testing.T) {
 		}
 	}
 }
-func TestPositiveResourceShouldBe(t *testing.T) {
-	var (
-		err                 error
-		g                   = gomega.NewWithT(t)
-		dynScheme           = runtime.NewScheme()
-		fakeDynamicClient   = fakeDynamic.NewSimpleDynamicClient(dynScheme)
-		fakeDiscovery       = fakeDiscovery.FakeDiscovery{}
-		fakeClient          *fake.Clientset
-		testResource        *unstructured.Unstructured
-		createdReactionFunc = func(action kTesting.Action) (handled bool, ret runtime.Object, err error) {
-			return true, testResource, nil
-		}
-		deletedReactionFunc = func(action kTesting.Action) (handled bool, ret runtime.Object, err error) {
-			return true, nil, kerrors.NewNotFound(schema.GroupResource{}, testResource.GetName())
-		}
-	)
-
-	const fileName = "test-resourcefile.yaml"
-
-	testResource, err = resourceFromYaml(fileName)
-	if err != nil {
-		t.Errorf("Failed getting the test resource from the file %v: %v", fileName, err)
-	}
-
-	fakeDiscovery.Fake = &fakeDynamicClient.Fake
-	fakeDiscovery.Resources = append(fakeDiscovery.Resources, newTestAPIResourceList(testResource.GetAPIVersion(), testResource.GetName(), testResource.GetKind()))
-
-	fakeDynamicClient.PrependReactor("get", "someResource", createdReactionFunc)
-
-	kc := Client{
-		DynamicInterface:   fakeDynamicClient,
-		DiscoveryInterface: &fakeDiscovery,
-		KubeInterface:      fakeClient,
-		FilesPath:          "../../test/templates",
-	}
-
-	err = kc.ResourceShouldBe(fileName, StateCreated)
-	g.Expect(err).ShouldNot(gomega.HaveOccurred())
-
-	fakeDiscovery.ReactionChain[0] = &kTesting.SimpleReactor{
-		Verb:     "get",
-		Resource: "someResource",
-		Reaction: deletedReactionFunc,
-	}
-
-	err = kc.ResourceShouldBe(fileName, StateDeleted)
-	g.Expect(err).ShouldNot(gomega.HaveOccurred())
-}
-
-func TestPositiveResourceShouldConvergeToSelector(t *testing.T) {
-
-	var (
-		err               error
-		g                 = gomega.NewWithT(t)
-		fakeDynamicClient = fakeDynamic.NewSimpleDynamicClient(runtime.NewScheme())
-		fakeDiscovery     = fakeDiscovery.FakeDiscovery{}
-		fakeClient        *fake.Clientset
-		testResource      *unstructured.Unstructured
-		getReactionFunc   = func(action kTesting.Action) (handled bool, ret runtime.Object, err error) {
-			return true, testResource, nil
-		}
-	)
-
-	const (
-		fileName = "test-resourcefile.yaml"
-		selector = ".metadata.labels.someTestKey=someTestValue"
-	)
-
-	testResource, err = resourceFromYaml(fileName)
-	if err != nil {
-		t.Errorf("Failed getting the test resource from the file %v: %v", fileName, err)
-	}
-
-	fakeDiscovery.Fake = &fakeDynamicClient.Fake
-	fakeDiscovery.Resources = append(fakeDiscovery.Resources, newTestAPIResourceList(testResource.GetAPIVersion(), testResource.GetName(), testResource.GetKind()))
-
-	fakeDynamicClient.PrependReactor("get", "someResource", getReactionFunc)
-
-	kc := Client{
-		DynamicInterface:   fakeDynamicClient,
-		DiscoveryInterface: &fakeDiscovery,
-		KubeInterface:      fakeClient,
-		FilesPath:          "../../test/templates",
-	}
-
-	err = kc.ResourceShouldConvergeToSelector(fileName, selector)
-	g.Expect(err).ShouldNot(gomega.HaveOccurred())
-}
-
-func TestPositiveResourceConditionShouldBe(t *testing.T) {
-
-	var (
-		err               error
-		g                 = gomega.NewWithT(t)
-		fakeDynamicClient = fakeDynamic.NewSimpleDynamicClient(runtime.NewScheme())
-		fakeDiscovery     = fakeDiscovery.FakeDiscovery{}
-		fakeClient        *fake.Clientset
-		testResource      *unstructured.Unstructured
-		getReactionFunc   = func(action kTesting.Action) (handled bool, ret runtime.Object, err error) {
-			return true, testResource, nil
-		}
-	)
-
-	const (
-		fileName            = "test-resourcefile.yaml"
-		testConditionType   = "someConditionType"
-		testConditionStatus = "true"
-	)
-
-	testResource, err = resourceFromYaml(fileName)
-	if err != nil {
-		t.Errorf("Failed getting the test resource from the file %v: %v", fileName, err)
-	}
-
-	fakeDiscovery.Fake = &fakeDynamicClient.Fake
-	fakeDiscovery.Resources = append(fakeDiscovery.Resources, newTestAPIResourceList(testResource.GetAPIVersion(), testResource.GetName(), testResource.GetKind()))
-
-	fakeDynamicClient.PrependReactor("get", "someResource", getReactionFunc)
-
-	kc := Client{
-		DynamicInterface:   fakeDynamicClient,
-		DiscoveryInterface: &fakeDiscovery,
-		KubeInterface:      fakeClient,
-		FilesPath:          "../../test/templates",
-	}
-
-	err = kc.ResourceConditionShouldBe(fileName, testConditionType, testConditionStatus)
-	g.Expect(err).ShouldNot(gomega.HaveOccurred())
-}
-
-func TestPositiveUpdateResourceWithField(t *testing.T) {
-
-	const (
-		fileName           = "test-resourcefile.yaml"
-		testUpdateKeyChain = ".metadata.labels.testUpdateKey"
-		testUpdateKey      = "testUpdateKey"
-		testUpdateValue    = "testUpdateValue"
-	)
-
-	var (
-		err               error
-		g                 = gomega.NewWithT(t)
-		fakeDynamicClient = fakeDynamic.NewSimpleDynamicClient(runtime.NewScheme())
-		fakeDiscovery     = fakeDiscovery.FakeDiscovery{}
-		fakeClient        *fake.Clientset
-		testResource      *unstructured.Unstructured
-		getReactionFunc   = func(action kTesting.Action) (handled bool, ret runtime.Object, err error) {
-			return true, testResource, nil
-		}
-		updateReactionFunc = func(action kTesting.Action) (handled bool, ret runtime.Object, err error) {
-			addLabel(testResource, testUpdateKey, testUpdateValue)
-			return true, testResource, nil
-		}
-	)
-
-	testResource, err = resourceFromYaml(fileName)
-	if err != nil {
-		t.Errorf("Failed getting the test resource from the file %v: %v", fileName, err)
-	}
-
-	fakeDiscovery.Fake = &fakeDynamicClient.Fake
-	fakeDiscovery.Resources = append(fakeDiscovery.Resources, newTestAPIResourceList(testResource.GetAPIVersion(), testResource.GetName(), testResource.GetKind()))
-
-	fakeDynamicClient.PrependReactor("get", "someResource", getReactionFunc)
-	fakeDynamicClient.PrependReactor("update", "someResource", updateReactionFunc)
-
-	kc := Client{
-		DynamicInterface:   fakeDynamicClient,
-		DiscoveryInterface: &fakeDiscovery,
-		KubeInterface:      fakeClient,
-		FilesPath:          "../../test/templates",
-	}
-
-	err = kc.UpdateResourceWithField(fileName, testUpdateKeyChain, testUpdateValue)
-	g.Expect(err).ShouldNot(gomega.HaveOccurred())
-
-	expectedLabelValue, found, err := unstructured.NestedString(testResource.UnstructuredContent(), "metadata", "labels", "testUpdateKey")
-	g.Expect(err).ShouldNot(gomega.HaveOccurred())
-	g.Expect(found).To(gomega.BeTrue())
-	g.Expect(expectedLabelValue).To(gomega.Equal(testUpdateValue))
-}
 
 func TestResourceInNamespace(t *testing.T) {
 	var (
@@ -421,7 +202,7 @@ func TestResourceInNamespace(t *testing.T) {
 		},
 	}
 
-	kc := Client{
+	kc := ClientSet{
 		KubeInterface:      fakeKubeClient,
 		DynamicInterface:   fakeDynamicClient,
 		DiscoveryInterface: fakeDiscovery,
@@ -478,7 +259,7 @@ func TestScaleDeployment(t *testing.T) {
 		replicaCount   = int32(1)
 	)
 
-	kc := Client{
+	kc := ClientSet{
 		KubeInterface: fakeKubeClient,
 	}
 
@@ -503,25 +284,7 @@ func TestScaleDeployment(t *testing.T) {
 	g.Expect(s.Spec.Replicas).To(gomega.Equal(int32(2)))
 }
 
-func resourceFromYaml(resourceFileName string) (*unstructured.Unstructured, error) {
-
-	resourcePath := filepath.Join("../../test/templates", resourceFileName)
-	d, err := ioutil.ReadFile(resourcePath)
-	if err != nil {
-		return nil, err
-	}
-	return resourceFromBytes(d)
-}
-func resourceFromBytes(bytes []byte) (*unstructured.Unstructured, error) {
-	resource := &unstructured.Unstructured{}
-	dec := serializer.NewDecodingSerializer(unstructured.UnstructuredJSONScheme)
-	_, _, err := dec.Decode(bytes, nil, resource)
-	if err != nil {
-		return nil, err
-	}
-
-	return resource, nil
-}
+// TODO: this is implemented twice, maybe have a test helper pkg?
 func newTestAPIResourceList(apiVersion, name, kind string) *metav1.APIResourceList {
 	return &metav1.APIResourceList{
 		GroupVersion: apiVersion,
@@ -535,17 +298,6 @@ func newTestAPIResourceList(apiVersion, name, kind string) *metav1.APIResourceLi
 	}
 }
 
-func addLabel(in *unstructured.Unstructured, key, value string) {
-	labels, _, _ := unstructured.NestedMap(in.Object, "metadata", "labels")
-
-	labels[key] = value
-
-	err := unstructured.SetNestedMap(in.Object, labels, "metadata", "labels")
-	if err != nil {
-		log.Errorf("Failed adding label %v=%v to the resource %v: %v", key, value, in.GetName(), err)
-	}
-}
-
 func TestClusterRoleAndBindingIsFound(t *testing.T) {
 	var (
 		err            error
@@ -553,7 +305,7 @@ func TestClusterRoleAndBindingIsFound(t *testing.T) {
 		fakeKubeClient = fake.NewSimpleClientset()
 	)
 
-	kc := Client{
+	kc := ClientSet{
 		KubeInterface: fakeKubeClient,
 	}
 
@@ -595,161 +347,7 @@ func TestClusterRoleAndBindingIsFound(t *testing.T) {
 	}
 }
 
-func Test_unstructuredResourceOperation(t *testing.T) {
-	type clientFields struct {
-		DynamicInterface dynamic.Interface
-	}
-	type funcArgs struct {
-		operation            string
-		ns                   string
-		unstructuredResource util.K8sUnstructuredResource
-	}
-
-	resourceNoNs, err := resourceFromYaml("../../test/templates/resource-without-namespace.yaml")
-	if err != nil {
-		t.Errorf(err.Error())
-	}
-
-	resourceNs, err := resourceFromYaml("../../test/templates/resource-with-namespace.yaml")
-	if err != nil {
-		t.Errorf(err.Error())
-	}
-
-	resourceNoNsUpdate, err := resourceFromYaml("../../test/templates/resource-without-namespace-update.yaml")
-	if err != nil {
-		t.Errorf(err.Error())
-	}
-
-	resourceNsUpdate, err := resourceFromYaml("../../test/templates/resource-with-namespace-update.yaml")
-	if err != nil {
-		t.Errorf(err.Error())
-	}
-
-	dynScheme := runtime.NewScheme()
-	fakeDynamicClient := fakeDynamic.NewSimpleDynamicClient(dynScheme)
-
-	tests := []struct {
-		name         string
-		clientFields clientFields
-		funcArgs     funcArgs
-		wantErr      bool
-	}{
-		{
-			name: "Resource create succeeds when namespace is configurable",
-			clientFields: clientFields{
-				DynamicInterface: fakeDynamicClient,
-			},
-			funcArgs: funcArgs{
-				operation: "create",
-				ns:        "test-namespace",
-				unstructuredResource: util.K8sUnstructuredResource{
-					GVR:      &meta.RESTMapping{},
-					Resource: resourceNoNs,
-				},
-			},
-			wantErr: false,
-		},
-		{
-			name: "Resource update succeeds when namespace is configurable",
-			clientFields: clientFields{
-				DynamicInterface: fakeDynamicClient,
-			},
-			funcArgs: funcArgs{
-				operation: "update",
-				ns:        "test-namespace",
-				unstructuredResource: util.K8sUnstructuredResource{
-					GVR:      &meta.RESTMapping{},
-					Resource: resourceNoNsUpdate,
-				},
-			},
-			wantErr: false,
-		},
-		{
-			name: "Resource delete succeeds when namespace is configurable",
-			clientFields: clientFields{
-				DynamicInterface: fakeDynamicClient,
-			},
-			funcArgs: funcArgs{
-				operation: "delete",
-				ns:        "test-namespace",
-				unstructuredResource: util.K8sUnstructuredResource{
-					GVR:      &meta.RESTMapping{},
-					Resource: resourceNoNs,
-				},
-			},
-			wantErr: false,
-		},
-		{
-			name: "Resource create succeeds when namespace in YAML",
-			clientFields: clientFields{
-				DynamicInterface: fakeDynamicClient,
-			},
-			funcArgs: funcArgs{
-				operation: "create",
-				unstructuredResource: util.K8sUnstructuredResource{
-					GVR:      &meta.RESTMapping{},
-					Resource: resourceNs,
-				},
-			},
-			wantErr: false,
-		},
-		{
-			name: "Resource update succeeds when namespace is in YAML",
-			clientFields: clientFields{
-				DynamicInterface: fakeDynamicClient,
-			},
-			funcArgs: funcArgs{
-				operation: "update",
-				unstructuredResource: util.K8sUnstructuredResource{
-					GVR:      &meta.RESTMapping{},
-					Resource: resourceNsUpdate,
-				},
-			},
-			wantErr: false,
-		},
-		{
-			name: "Resource create fails when namespace configured and in YAML",
-			clientFields: clientFields{
-				DynamicInterface: fakeDynamicClient,
-			},
-			funcArgs: funcArgs{
-				operation: "create",
-				ns:        "override-ns",
-				unstructuredResource: util.K8sUnstructuredResource{
-					GVR:      &meta.RESTMapping{},
-					Resource: resourceNs,
-				},
-			},
-			wantErr: true,
-		},
-		{
-			name: "Unsupported operation produces error",
-			clientFields: clientFields{
-				DynamicInterface: fakeDynamicClient,
-			},
-			funcArgs: funcArgs{
-				operation: "invalid",
-				unstructuredResource: util.K8sUnstructuredResource{
-					GVR:      &meta.RESTMapping{},
-					Resource: resourceNs,
-				},
-			},
-			wantErr: true,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			kc := &Client{
-				DynamicInterface: tt.clientFields.DynamicInterface,
-			}
-			if err := kc.unstructuredResourceOperation(tt.funcArgs.operation, tt.funcArgs.ns, tt.funcArgs.unstructuredResource); (err != nil) != tt.wantErr {
-				t.Errorf("Client.unstructuredResourceOperation() error = %v, wantErr %v", err, tt.wantErr)
-			}
-		})
-	}
-}
-
-func Test_ThePodsInNamespaceWithSelectorShouldHaveLabels(t *testing.T) {
+func Test_PodsInNamespaceWithSelectorShouldHaveLabels(t *testing.T) {
 	ns := v1.Namespace{ObjectMeta: metav1.ObjectMeta{Name: "foo"}}
 	podWithLabels1 := v1.Pod{
 		ObjectMeta: metav1.ObjectMeta{
@@ -847,12 +445,12 @@ func Test_ThePodsInNamespaceWithSelectorShouldHaveLabels(t *testing.T) {
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			kc := &Client{
+			kc := &ClientSet{
 				KubeInterface:      tt.fields.KubeInterface,
 				DynamicInterface:   tt.fields.DynamicInterface,
 				DiscoveryInterface: tt.fields.DiscoveryInterface,
 			}
-			if err := kc.ThePodsInNamespaceWithSelectorShouldHaveLabels(tt.args.namespace, tt.args.selector, tt.args.labels); (err != nil) != tt.wantErr {
+			if err := kc.PodsInNamespaceWithSelectorShouldHaveLabels(tt.args.namespace, tt.args.selector, tt.args.labels); (err != nil) != tt.wantErr {
 				t.Errorf("ThePodsInNamespaceWithSelectorShouldHaveLabels() error = %v, wantErr %v", err, tt.wantErr)
 			}
 		})
